Patterns alternate: <a b>, and it cost no runtime state
'<a b>' plays a different member each bar. Angle brackets pick one, square
brackets subdivide, and they compose in either order:

    >kick x...<x .>...     a hit on the two, every other bar
    >hat  <x.x. xxxx>      swap a whole group, bar to bar
    >bass 0...<3 5>...     the value changes
    >snare <x%15 x%90>     and so do the odds
    >disc  <9 3>           pictures alternate too, because a lane is a lane

THE IMPLEMENTATION IS THE INTERESTING PART, and it is cheaper than planned.
docs/MAP.md estimated a per-slot table of alternatives indexed by a bar counter in
the fire path, about 128 bytes a lane. The right answer was the one nesting
already uses: expand at compile time. Lay the pattern down once per cycle with the
group resolved differently each time, and the lane's ordinary wrap does the
alternation - a sixteen-step lane alternating two ways is simply a thirty-two slot
lane, on the same flat bitmask at the same uniform rate. No variant table, no
cycle counter, nothing new in the realtime path that could be late.

The budget moved to pay for it: SEQ_MAX_STEPS is 64 and the four step bitmasks are
64-bit. Eight bytes a lane, and it lets a sixteen-step lane alternate three ways
or nest and alternate together. Every bit operation on those masks had to widen
with them; a missed '1u <<' would have silently lost the top thirty-two slots.

'%' travels with the alternative, which is the property that makes this worth
having rather than merely clever: a member is just an item, so 'x%15<3%20 5%80>'
keeps each alternative's own odds and '<x x%0>' is certain one bar and never the
next. Verified per slot rather than assumed.

ONE BUG WORTH RECORDING. '<[xx] x>' first compiled with the group given a single
slot, silently dropping half of it - the pattern compiled, played, and was quietly
wrong, which is the worst way for a subdivision to fail. The rule "how wide is one
item" had been written twice, once in seq_pattern_span and once in
seq_pattern_walk's top-level loop, and only one copy learned about '<>'. Two
copies of one rule is precisely the drift this header exists to prevent, and it
had drifted inside itself. There is one copy now, seq_pattern_item_span, and both
callers use it.

Tested on hardware at 140 bpm with seven lanes, music and picture alternating
together, and the odds reported back as 15,90.

+++ firmware/components/seq/seq.c
@@ -448,14 +448,14 @@ static void fire_lanes(uint32_t tick)
                 continue;
             }
             s_emitting = l->name;
-            if (!(l->mask & (1u << s))) {
+            if (!(l->mask & (1ull << s))) {
                 continue;
             }
         }
         /* '?' - maybe. Half, because half is the only ratio that needs no
          * number after it, and a number after it would be the start of the
          * syntax this instrument is trying not to have. */
-        if (l->route[0] == '\0' && (l->chance & (1u << s))) {
+        if (l->route[0] == '\0' && (l->chance & (1ull << s))) {
             /* '?' alone is half; '?[15]' is fifteen per cent. Half is the
              * default because it is the only ratio that needs no number, and
              * the bracket is there for when the player wants a different one
@@ -471,8 +471,8 @@ static void fire_lanes(uint32_t tick)
          * numbers, so setting a lane quiet keeps its accents in proportion
          * instead of flattening the whole pattern against a ceiling. */
         int vel = l->vel;
-        if (l->accent & (1u << s)) { vel = vel + (127 - vel) * 3 / 4; }
-        if (l->ghost  & (1u << s)) { vel = vel / 3; }
+        if (l->accent & (1ull << s)) { vel = vel + (127 - vel) * 3 / 4; }
+        if (l->ghost  & (1ull << s)) { vel = vel / 3; }
         if (vel < 1)   { vel = 1; }
         if (vel > 127) { vel = 127; }
 
@@ -791,7 +791,7 @@ esp_err_t seq_lane(const char *name, const char *steps)
         return ESP_ERR_INVALID_SIZE;
     }
 
-    uint32_t mask = 0, accent = 0, ghost = 0, chance = 0;
+    uint64_t mask = 0, accent = 0, ghost = 0, chance = 0;
     char chr[SEQ_MAX_STEPS];
     uint8_t deg[SEQ_MAX_STEPS];
     uint8_t prob[SEQ_MAX_STEPS];
@@ -812,11 +812,11 @@ esp_err_t seq_lane(const char *name, const char *steps)
         if (ch == '.' || ch == '-' || ch == '_') {
             continue;
         }
-        mask |= (1u << i);
+        mask |= (1ull << i);
         chr[i] = ch;
-        if (ch == 'X') { accent |= (1u << i); }
-        if (ch == ',') { ghost  |= (1u << i); }
-        if (ch == '?') { chance |= (1u << i); }
+        if (ch == 'X') { accent |= (1ull << i); }
+        if (ch == ',') { ghost  |= (1ull << i); }
+        if (ch == '?') { chance |= (1ull << i); }
         if (ch >= '0' && ch <= '9') { deg[i] = (uint8_t)(ch - '0'); }
 
         /* '%NN' is a parameter on this step, and it implies maybe. 0 means
@@ -825,7 +825,7 @@ esp_err_t seq_lane(const char *name, const char *steps)
         const int v = seq_pattern_param(steps + off + 1);
         if (v >= 0 && v <= 100) {
             prob[i] = (uint8_t)v;
-            chance |= (1u << i);
+            chance |= (1ull << i);
         }
     }
